Update capabilities.md
I've changed the last three from nouns to verbs, though I did not yet change the pointer they are going to. If approved, we should change those names too. 

I am proposing that we combine (recombine) Allocate and Report into Understand Fully Loaded Cost or Understand Spend and Usage. I think both Allocate and Report could be Functions under that Capability. I'm not sure how to make that broad a change though. 

I also put in a brief description of what Capabilities are. I'm not certain I agree that Capabilities are the right term for this. I think these are Functional Categories maybe but I'm willing to go with what's here for now. I'm not sure what differentiates Functions and Capabilities in this sense if Capability is just the word we use for a group of functions. What the Framework calls Functions we call Capabilities in the FOCP training, so whatever we decide should update everywhere.